requirements.txt
torch
torchvision
flask
numpy
pillow
opencv-python
matplotlib